requirements.txt
paho-mqtt
influxdb
tcp-latency
psutil
